refactor extensions folder location
