+++ extensions/index.ts
@@ -1,2 +1,3 @@
 export * from './app-routes';
 export * from './mongodb-server-dev';
+export * from './media-server';
